authentication

+++ app/api/tasks/[taskId]/route.ts
@@ -1,6 +1,8 @@
 import prisma from "@/prisma/client";
 import { revalidatePath } from "next/cache";
+import { headers } from "next/dist/client/components/headers";
 import { NextRequest, NextResponse } from "next/server";
+import jwt from "jsonwebtoken";
 
 interface ContextParams {
   params: {
@@ -12,6 +14,12 @@ export async function PUT(
   _req: NextRequest,
   { params: { taskId } }: ContextParams
 ) {
+  const headersList = headers();
+  const token = headersList.get("Authorization");
+  if (!token) throw Error("No token provided");
+  if (!process.env.JWT_SECRET) throw Error("No secret found");
+  jwt.verify(token, process.env.JWT_SECRET);
+
   const task = await prisma.task.findUniqueOrThrow({
     where: { id: parseInt(taskId) },
   });
@@ -37,6 +45,12 @@ export async function DELETE(
   _req: NextRequest,
   { params: { taskId } }: ContextParams
 ) {
+  const headersList = headers();
+  const token = headersList.get("Authorization");
+  if (!token) throw Error("No token provided");
+  if (!process.env.JWT_SECRET) throw Error("No secret found");
+  jwt.verify(token, process.env.JWT_SECRET);
+
   revalidatePath("/");
   return NextResponse.json(
     await prisma.task.delete({
